Moved write tests to formats
